requirements.txt
torch>=2.0.0
gymnasium[atari,accept-rom-license]>=1.0.0
ale-py>=0.9.0
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
pyyaml>=6.0
opencv-python>=4.8.0
